requirements.txt
selenium
pytest
allure-pytest
requests
webdriver-manager
